+++ docs/CHANGELOG.md
@@ -6,3 +6,4 @@
 + Fix empty polygon with ST_TESSELLATE
 + Add ST_ConcaveHull function
 + Improve ST_Graph function to keep a list of columns for the edges
++ Add ST_LineSubstring and ST_LineInterpolatePoint functions